refactor: remove superseded OPF sidecar export (#1373)
The OPF sidecar export (F5.8) wrote an external metadata.opf next to a
book's EPUB, but external readers and kepubify only ever read metadata
baked inside the .epub container, so the sidecar never surfaced a
user's edits on a device. Overrides now bake directly into exports via
epub_rewrite (#1372), making the sidecar redundant and misleading.

Removes db::opf_export (module + tests), the
POST /api/ebooks/{uuid}/export-opf REST route and rpc_export_opf
server function, the two frontend export_opf data-layer wrappers, the
"Export to OPF" sidebar button on the metadata-edit page, and the
OpfExportResult wire DTO. The shared render_effective_metadata /
xml_escape metadata renderer moves into epub_rewrite::opf, its one
remaining consumer. Updates docs/architecture.md and the affected
Playwright spec.
