+++ db/src/epub_rewrite/opf.rs
@@ -6,21 +6,21 @@
 //! refinements, and the whole manifest/spine — is preserved untouched, so the
 //! EPUB's identity and structure survive.
 
+use std::fmt::Write as _;
+
 use anyhow::Context;
 use omnibus_shared::EbookMetadata;
 use quick_xml::events::{BytesStart, Event};
 use quick_xml::reader::Reader;
 use quick_xml::writer::Writer;
 
-use crate::opf_export::render_effective_metadata;
-
 /// Descriptive `<dc:*>` local names we own: their originals are dropped and a
 /// fresh copy is emitted from the effective metadata. Everything not listed
 /// (`identifier`, `rights`, `source`, `type`, …) is preserved verbatim.
 ///
 /// `contributor` is dropped alongside `creator` because the effective
-/// `creators` vec already folds contributors in (mirroring `render_opf`), so
-/// re-emitting them as `<dc:creator>` would otherwise duplicate them.
+/// `creators` vec already folds contributors in, so re-emitting them as
+/// `<dc:creator>` would otherwise duplicate them.
 const MANAGED_DC: &[&[u8]] = &[
     b"title",
     b"creator",
@@ -243,4 +243,109 @@ fn local_name(qname: &[u8]) -> &[u8] {
         Some(i) => &qname[i + 1..],
         None => qname,
     }
+}
+
+/// Render the descriptive `<metadata>` children this rewrite regenerates from
+/// a book's effective metadata — title, creators, description, publisher,
+/// date, language, subjects, and the calibre series pair — as a single
+/// four-space-indented XML fragment, one element per line.
+///
+/// Contributors were merged into `creators` at parse time and the schema
+/// can't tell them apart, so every creator re-exports as `<dc:creator>`.
+/// Description is ammonia-sanitized HTML; escaping it once as a text node
+/// means a consumer that unescapes gets the markup back.
+fn render_effective_metadata(book: &EbookMetadata) -> String {
+    let mut out = String::new();
+    push_title_and_creators(&mut out, book);
+    push_dc(&mut out, "description", book.description.as_deref());
+    push_dc(&mut out, "publisher", book.publisher.as_deref());
+    push_dc(&mut out, "date", book.published.as_deref());
+    push_dc(&mut out, "language", book.language.as_deref());
+    push_subjects(&mut out, book);
+    push_series_metas(&mut out, book);
+    out
+}
+
+/// Emit `dc:title` (mandatory in OPF 2.0, falling back to the filename the
+/// way the UI does for untitled books) and one `dc:creator` per contributor.
+fn push_title_and_creators(out: &mut String, book: &EbookMetadata) {
+    let title = book
+        .title
+        .as_deref()
+        .filter(|s| !s.is_empty())
+        .unwrap_or(&book.filename);
+    let _ = writeln!(out, "    <dc:title>{}</dc:title>", xml_escape(title));
+
+    for creator in &book.creators {
+        let mut attrs = String::new();
+        if let Some(role) = creator.role.as_deref().filter(|s| !s.is_empty()) {
+            let _ = write!(attrs, " opf:role=\"{}\"", xml_escape(role));
+        }
+        if let Some(file_as) = creator.file_as.as_deref().filter(|s| !s.is_empty()) {
+            let _ = write!(attrs, " opf:file-as=\"{}\"", xml_escape(file_as));
+        }
+        let _ = writeln!(
+            out,
+            "    <dc:creator{}>{}</dc:creator>",
+            attrs,
+            xml_escape(&creator.name)
+        );
+    }
+}
+
+/// Emit one `dc:subject` per subject tag.
+fn push_subjects(out: &mut String, book: &EbookMetadata) {
+    for subject in &book.subjects {
+        let _ = writeln!(out, "    <dc:subject>{}</dc:subject>", xml_escape(subject));
+    }
+}
+
+/// Emit the Calibre EPUB2 series `<meta>` pair — the same keys the parser
+/// falls back to, so a rewrite→rescan round-trips.
+fn push_series_metas(out: &mut String, book: &EbookMetadata) {
+    if let Some(series) = book.series.as_deref().filter(|s| !s.is_empty()) {
+        let _ = writeln!(
+            out,
+            "    <meta name=\"calibre:series\" content=\"{}\"/>",
+            xml_escape(series)
+        );
+    }
+    if let Some(index) = book.series_index.as_deref().filter(|s| !s.is_empty()) {
+        let _ = writeln!(
+            out,
+            "    <meta name=\"calibre:series_index\" content=\"{}\"/>",
+            xml_escape(index)
+        );
+    }
+}
+
+/// Emit `<dc:{tag}>{value}</dc:{tag}>` when `value` is present and non-empty.
+fn push_dc(out: &mut String, tag: &str, value: Option<&str>) {
+    if let Some(v) = value.filter(|s| !s.is_empty()) {
+        let _ = writeln!(out, "    <dc:{tag}>{}</dc:{tag}>", xml_escape(v));
+    }
+}
+
+/// Escape the five XML metacharacters and strip XML-illegal control
+/// characters. Used for both text nodes and attribute values —
+/// over-escaping `"`/`'` in a text node is still valid XML.
+fn xml_escape(s: &str) -> String {
+    let mut out = String::with_capacity(s.len());
+    for c in s.chars().filter(|c| !is_xml_illegal_control_char(*c)) {
+        match c {
+            '&' => out.push_str("&amp;"),
+            '<' => out.push_str("&lt;"),
+            '>' => out.push_str("&gt;"),
+            '"' => out.push_str("&quot;"),
+            '\'' => out.push_str("&apos;"),
+            _ => out.push(c),
+        }
+    }
+    out
+}
+
+/// True for control characters that are valid UTF-8 but illegal in XML 1.0
+/// text (everything below 0x20 except tab, LF, and CR).
+fn is_xml_illegal_control_char(c: char) -> bool {
+    matches!(c, '\u{0}'..='\u{8}' | '\u{B}'..='\u{C}' | '\u{E}'..='\u{1F}')
 }
