+++ frontend/src/pages/metadata_edit/sidebar.rs
@@ -1,14 +1,13 @@
 //! Sticky right-column sidebar for the metadata edit page: cover
 //! upload/revert (delegated to `cover_editor::CoverEditor`), identifiers,
-//! the OPF-export card, and the override-active card. Full-override revert
-//! bubbles to the parent `MetadataEditForm`; cover-only revert stays local,
-//! updating `live_book` so the "Override active" card tracks it.
+//! and the override-active card. Full-override revert bubbles to the parent
+//! `MetadataEditForm`; cover-only revert stays local, updating `live_book` so
+//! the "Override active" card tracks it.
 
 use dioxus::prelude::*;
-use omnibus_shared::{EbookMetadata, OpfExportResult};
+use omnibus_shared::EbookMetadata;
 
 use super::cover_editor::CoverEditor;
-use crate::{data, use_server_url};
 
 /// Cover preview + identifiers + override-status sidebar.
 #[component]
@@ -18,7 +17,6 @@ pub(super) fn Sidebar(
     on_revert: EventHandler<()>,
 ) -> Element {
     let uuid = book.unique_identifier.clone().unwrap_or_default();
-    let export_uuid = uuid.clone();
     let identifiers = book.identifiers.clone();
     // Tracks the merged book returned by a cover upload/revert so the
     // "Override active" card reflects a cover-only change immediately —
@@ -56,8 +54,6 @@ pub(super) fn Sidebar(
                 }
             }
 
-            ExportCard { uuid: export_uuid, saving }
-
             // Override status
             if has_override {
                 div { class: "card me-sidebar-card",
@@ -76,106 +72,5 @@ pub(super) fn Sidebar(
                 }
             }
         }
-    }
-}
-
-/// "Export to OPF" card: writes the book's merged metadata to its
-/// `metadata.opf` sidecar and reports the resulting path / backup status.
-/// Not dirty-dependent, so it owns its own state instead of touching the
-/// parent form's save flow.
-#[component]
-fn ExportCard(uuid: String, saving: Signal<bool>) -> Element {
-    let server_url = use_server_url();
-    let mut exporting = use_signal(|| false);
-    let mut status: Signal<Option<Result<OpfExportResult, String>>> = use_signal(|| None);
-
-    let on_export = move |_| {
-        let uuid = uuid.clone();
-        let server_url = server_url.clone();
-        exporting.set(true);
-        status.set(None);
-        spawn(async move {
-            let result = data::export_opf(&server_url, &uuid)
-                .await
-                .map_err(|e| export_error_message(&e));
-            status.set(Some(result));
-            exporting.set(false);
-        });
-    };
-
-    rsx! {
-        div { class: "card me-sidebar-card",
-            div { class: "label", style: "margin-bottom: 8px;", "Export" }
-            p { class: "mono", style: "font-size: 11px; color: var(--ink-2);",
-                "Write a metadata.opf sidecar next to the book's EPUB with the current (scanned + override) metadata."
-            }
-            button {
-                class: "btn ghost sm",
-                style: "margin-top: 10px; width: 100%; justify-content: center;",
-                "data-testid": "me-export-opf",
-                disabled: exporting() || saving(),
-                onclick: on_export,
-                if exporting() { "Exporting\u{2026}" } else { "Export to OPF" }
-            }
-            match status() {
-                Some(Ok(result)) => rsx! {
-                    p {
-                        class: "mono",
-                        "data-testid": "me-export-status",
-                        style: "font-size: 11px; color: var(--ink-2); margin-top: 8px;",
-                        "Wrote {result.path}"
-                        if result.backed_up {
-                            " (previous file backed up to metadata.opf.bak)"
-                        }
-                    }
-                },
-                Some(Err(msg)) => rsx! {
-                    p {
-                        class: "mono",
-                        "data-testid": "me-export-status",
-                        style: "font-size: 11px; color: var(--bad); margin-top: 8px;",
-                        "{msg}"
-                    }
-                },
-                None => rsx! {},
-            }
-        }
-    }
-}
-
-/// Build the status-line message for a failed export, splicing in the
-/// server's actual diagnostic text (via [`data::server_error_message`])
-/// instead of `DataError`'s status-only `Display` — a 409/403 should read
-/// as its real reason, not "server returned {status}".
-fn export_error_message(err: &data::DataError) -> String {
-    format!("Export failed: {}", data::server_error_message(err))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn export_error_message_surfaces_the_servers_diagnostic_body() {
-        let err = data::DataError::Http {
-            status: 409,
-            body: "book has no EPUB file to export next to".into(),
-        };
-        assert_eq!(
-            export_error_message(&err),
-            "Export failed: 409: book has no EPUB file to export next to"
-        );
-    }
-
-    #[test]
-    fn export_error_message_falls_back_to_display_when_the_body_is_empty() {
-        let err = data::DataError::Http {
-            status: 500,
-            body: String::new(),
-        };
-        assert_eq!(
-            export_error_message(&err),
-            "Export failed: server returned 500"
-        );
     }
 }
