+++ shared/src/ebook/mod.rs
@@ -5,7 +5,6 @@
 //! layer persisted in `metadata_overrides.overrides` and merged on read.
 //! Re-exports flatten so callers keep `omnibus_shared::EbookMetadata` etc.
 
-mod export;
 mod metadata;
 mod overrides;
 mod validator;
@@ -13,7 +12,6 @@ mod validator;
 #[cfg(test)]
 mod tests;
 
-pub use export::OpfExportResult;
 pub use metadata::{BookFileInfo, Contributor, EbookMetadata, Identifier};
 pub use overrides::MetadataOverrides;
 pub use validator::{
